Change workflow in EGLSync.
EGLSync is enabled when a correct manifest or wine prefix
folder is specified.
If the path is removed, sync is disabled. (--unlink)
Automatic sync is controlled by a chechbox. It doesn't sync
on first activation for now.
`Sync once` is removed because it can be serviced
by `Select all` -> `Import`/`Export`

Still work in progress.