GameWidget: work-around crash in `eventFilter`
For an unknown reason presently, the eventFilter is called with
a QRunnable as the second argument instead of a QEvent. This causes
a crash because a QRunnable doesn't have a `type` attribute. The target
object seems to be the launch button.

Work-around it by filtering out the event early if the second argument
is not a QEevent instance and log an error.