RareGameSlim: fix uploading save games when there aren't any on the remote
If there aren't any saves for the game on the cloud, the game's saves list
is empty, and the first check failed. As a result the method would
erroneously report that there aren't any save games. This caused the
upload/download button in the cloud saves dialog and widget would be
disabled.

Fix it by removing the check and creating a fake save to satisfy
`check_savegame_state`