ImageManager: test image data downloaded from Epic before using them.
In rare cases, image data from Epic might be invalid, resulting in an image with 0 x/y dimensions.
This later results in a ZeroDivisionError while trying to create the icon.

Work around it by testing the received data early and replacing it with the fallback image
if they do not have valid dimensions. Invalidate the image checksum for it to be redownloaded
at the next startup.

fixes: RareDevs/Rare#567