chore: remove `version_scheme` from `pyproject.toml`
Using `pip install .` or `python -m build` fails if this directive it
present. Both of these commands work correctly without the directive by
using a barebones `setup.py`.

On the other hand, removing this directive breaks
`python3 -m setuptools_scm --force-write-version-files` which is used
in the workflows to generate `rare/_version.py` before packaging. In the
case of the workflows though we already override the version using
`SETUPTOOLS_SCM_PRETEND_VERSION` so it shouldn't be an issue.