IndicatorLineEdit: add validation signal and refresh method
Add a `validationFinished(bool, str)` signal to indicate that is emitted
when the input validation has finished. The result contains if the input
was valid, and the reason is string form.

The `refresh()` method is meant to be use to refresh re-validate and refresh
the indicator upon request. This can be used when showing the widget to make
sure the existing input is valid.