Rare: update to use CloudSyncWorker
Because syncing is a network operation and verify/move are disk IO
operation, a second threadpool and queue were added to contain separate
them and queue the workers independently. One threadpool handles disk io
tasks and the other networking tasks with a maximum of 2 active threads
for each of them.