RunnerSettings: reduce memory usage caused by multiple copies of steam appids in the completer
The completer needs to use it's own copy of steam app ids which increases
the RAM usage by about 150MB. Using a bespoke item model doesn't work, as
altering the view of an ordered dictionary is a not efficient enough,
causing the UI to lock up. For that reason, we generate a reduce list of
completions based on the first three letters of the application's name
offer only those. The completer is replaced when the the view is shown
and destroyed when the view get hidden.